requirements.txt
Flask==3.1.2
gunicorn==23.0.0
python-dotenv==1.1.1
Flask-Cors==6.0.1
reportlab==4.4.1
